Implement CalculationService, fix returned types

+++ Application/Commands/SingleOperand/SingleOperandCommand.cs
@@ -2,7 +2,7 @@
 
 namespace Application.Commands.SingleOperand
 {
-    public class SingleOperandCommand : IRequest<int>
+    public class SingleOperandCommand : IRequest<long>
     {
         public int Value { get; set; }
         public string Operator { get; set; }
